It seems more appropriate to add first_part_of_celestial_coordinates to extract instead of fitting, since it mimics code from extract.Detection._physical coordinates.
